% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_matching.R
\name{calculate_l1_imbalance}
\alias{calculate_l1_imbalance}
\title{Calculate L1 Imbalance Measure}
\usage{
calculate_l1_imbalance(data, treatment_var, matching_variables)
}
\description{
Calculate L1 Imbalance Measure
}
\keyword{internal}
