% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_metrics.R
\name{calculate_transition_pattern_metrics}
\alias{calculate_transition_pattern_metrics}
\title{Calculate Transition Pattern Metrics}
\usage{
calculate_transition_pattern_metrics(
  data,
  id_column = "cf",
  period_column = "event_period",
  date_column = "inizio"
)
}
\arguments{
\item{data}{A data.table containing employment records with event identification}

\item{id_column}{Character. Name of person identifier column. Default: "cf"}

\item{period_column}{Character. Column indicating pre/post event period. Default: "event_period"}

\item{date_column}{Character. Name of date column. Default: "inizio"}
}
\value{
A data.table with transition pattern metrics:
  \item{cf}{Person identifier}
  \item{period}{Pre or post event period}
  \item{avg_time_to_next_job}{Average days from job end to next job start}
  \item{median_time_to_next_job}{Median days from job end to next job start}
  \item{job_search_success_rate}{Proportion of unemployment spells ending in employment}
  \item{contract_type_changes}{Number of contract type changes}
  \item{upward_transitions}{Number of transitions to better contract types}
  \item{downward_transitions}{Number of transitions to worse contract types}
  \item{transition_frequency}{Number of job transitions per year}
  \item{employment_continuity_index}{Measure of employment continuity}
}
\description{
Calculates transition pattern metrics including time to next job, contract type changes,
and transition frequency measures.
}
\examples{
\dontrun{
transition_metrics <- calculate_transition_pattern_metrics(
  data = event_data
)
}

}
