% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_matching.R
\name{cem_fallback_implementation}
\alias{cem_fallback_implementation}
\title{Fallback CEM Implementation (native R, no dependencies) with Person-Level Matching}
\usage{
cem_fallback_implementation(
  data,
  treatment_var,
  person_id_var,
  matching_variables,
  automatic_binning,
  cutpoints,
  k2k,
  control_ratio,
  keep_all,
  n_bins,
  verbose,
  original_data,
  aggregation_report
)
}
\description{
Fallback CEM Implementation (native R, no dependencies) with Person-Level Matching
}
\keyword{internal}
