% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_matching.R
\name{cem_native_implementation}
\alias{cem_native_implementation}
\title{Native CEM Implementation (using CEM package) with Person-Level Matching}
\usage{
cem_native_implementation(
  data,
  treatment_var,
  person_id_var,
  matching_variables,
  automatic_binning,
  cutpoints,
  k2k,
  control_ratio,
  keep_all,
  verbose,
  original_data,
  aggregation_report
)
}
\description{
Native CEM Implementation (using CEM package) with Person-Level Matching
}
\keyword{internal}
