% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggraph_transitions.R
\name{.add_circular_titles}
\alias{.add_circular_titles}
\title{Add Circular Plot Titles}
\usage{
.add_circular_titles(tg, title, subtitle, circular_type, node_order)
}
\arguments{
\item{tg}{tidygraph object}

\item{title}{Custom title}

\item{subtitle}{Custom subtitle}

\item{circular_type}{Type of circular layout}

\item{node_order}{Node ordering method}
}
\value{
ggplot2 labs layer
}
\description{
Internal helper to add titles for circular plots.
}
\keyword{internal}
