% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggraph_transitions.R
\name{.add_consolidation_info}
\alias{.add_consolidation_info}
\title{Add Consolidation Information to tidygraph}
\usage{
.add_consolidation_info(tg, transitions_data)
}
\arguments{
\item{tg}{tidygraph object}

\item{transitions_data}{Original transitions data with consolidation info}
}
\value{
Modified tidygraph object
}
\description{
Internal helper to add consolidation status to tidygraph edges.
}
\keyword{internal}
