% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggraph_transitions.R
\name{.add_heatmap_color_scale}
\alias{.add_heatmap_color_scale}
\title{Add Heatmap Color Scale}
\usage{
.add_heatmap_color_scale(scale_type, colors, accessibility_mode)
}
\arguments{
\item{scale_type}{Type of color scale}

\item{colors}{Color palette}

\item{accessibility_mode}{High contrast mode}
}
\value{
ggplot2 scale layer
}
\description{
Internal helper to add appropriate color scales to heatmap plots.
}
\keyword{internal}
