% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggraph_transitions.R
\name{.add_hierarchical_labels}
\alias{.add_hierarchical_labels}
\title{Add Hierarchical Labels}
\usage{
.add_hierarchical_labels(label_angle, accessibility_mode)
}
\arguments{
\item{label_angle}{Angle for labels}

\item{accessibility_mode}{High contrast mode}
}
\value{
ggraph label layer
}
\description{
Internal helper to add labels for hierarchical plots.
}
\keyword{internal}
