% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggraph_transitions.R
\name{.add_hierarchical_metrics}
\alias{.add_hierarchical_metrics}
\title{Add Hierarchical Metrics to Graph}
\usage{
.add_hierarchical_metrics(tg, hierarchy_type, root_nodes)
}
\arguments{
\item{tg}{tidygraph object}

\item{hierarchy_type}{Type of hierarchy}

\item{root_nodes}{Specified root nodes}
}
\value{
Modified tidygraph object with hierarchy info
}
\description{
Internal helper to add hierarchy-specific metrics to tidygraph object.
}
\keyword{internal}
