% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggraph_transitions.R
\name{.add_labels}
\alias{.add_labels}
\title{Add Network Labels}
\usage{
.add_labels(show_labels, label_size, label_repel, accessibility_mode)
}
\arguments{
\item{show_labels}{Show labels flag}

\item{label_size}{Size of labels}

\item{label_repel}{Use repelling labels}

\item{accessibility_mode}{High contrast mode}
}
\value{
ggraph label layer or NULL
}
\description{
Internal helper to add node labels to ggraph plots.
}
\keyword{internal}
