% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggraph_transitions.R
\name{.apply_circular_theme}
\alias{.apply_circular_theme}
\title{Apply Circular Theme}
\usage{
.apply_circular_theme(accessibility_mode)
}
\arguments{
\item{accessibility_mode}{High contrast mode}
}
\value{
ggplot2 theme layers
}
\description{
Internal helper to apply theme for circular plots.
}
\keyword{internal}
