% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggraph_transitions.R
\name{.convert_to_matrix}
\alias{.convert_to_matrix}
\title{Convert Data to Matrix Format}
\usage{
.convert_to_matrix(data, format)
}
\arguments{
\item{data}{Input transitions data}

\item{format}{Input format type}
}
\value{
Transition matrix
}
\description{
Internal helper to convert various input formats to matrix.
}
\keyword{internal}
