% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_reporting.R
\name{.create_academic_template}
\alias{.create_academic_template}
\title{Create Academic Template Content}
\usage{
.create_academic_template(params)
}
\description{
Create Academic Template Content
}
\keyword{internal}
