% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_rdd.R
\name{.create_density_plot}
\alias{.create_density_plot}
\title{Create Density Plot}
\usage{
.create_density_plot(dt, running_var, cutoff, bandwidth, colors)
}
\description{
Create Density Plot
}
\keyword{internal}
