% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_reporting.R
\name{.create_executive_template}
\alias{.create_executive_template}
\title{Create Executive Template Content}
\usage{
.create_executive_template(params)
}
\description{
Create Executive Template Content
}
\keyword{internal}
