% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_rdd.R
\name{.create_main_rdd_plot}
\alias{.create_main_rdd_plot}
\title{Create Main RDD Plot}
\usage{
.create_main_rdd_plot(
  dt,
  outcome_var,
  running_var,
  cutoff,
  bandwidth,
  bin_width,
  confidence_level,
  colors,
  show_points,
  point_alpha,
  bandwidth_lines,
  title_main,
  subtitle,
  rdd_results
)
}
\description{
Create Main RDD Plot
}
\keyword{internal}
