% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_reporting.R
\name{.create_policy_template}
\alias{.create_policy_template}
\title{Create Policy Template Content}
\usage{
.create_policy_template(params)
}
\description{
Create Policy Template Content
}
\keyword{internal}
