% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_reporting.R
\name{.create_rmd_template}
\alias{.create_rmd_template}
\title{Create R Markdown Template}
\usage{
.create_rmd_template(template, params)
}
\description{
Create R Markdown Template
}
\keyword{internal}
