% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_reporting.R
\name{.create_technical_template}
\alias{.create_technical_template}
\title{Create Technical Template Content}
\usage{
.create_technical_template(params)
}
\description{
Create Technical Template Content
}
\keyword{internal}
