% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_rdd.R
\name{.estimate_fuzzy_rdd}
\alias{.estimate_fuzzy_rdd}
\title{Estimate Fuzzy RDD}
\usage{
.estimate_fuzzy_rdd(
  dt,
  outcome,
  treatment_variable,
  running_variable,
  polynomial_order,
  kernel,
  bandwidth,
  bias_correction,
  robust_inference,
  cluster_variable,
  covariates,
  alpha
)
}
\description{
Estimate Fuzzy RDD
}
\keyword{internal}
