% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_rdd.R
\name{.estimate_sharp_rdd}
\alias{.estimate_sharp_rdd}
\title{Estimate Sharp RDD}
\usage{
.estimate_sharp_rdd(
  dt,
  outcome,
  running_variable,
  polynomial_order,
  kernel,
  bandwidth,
  bias_correction,
  robust_inference,
  cluster_variable,
  covariates,
  alpha
)
}
\description{
Estimate Sharp RDD
}
\keyword{internal}
