% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_reporting.R
\name{.format_balance_table}
\alias{.format_balance_table}
\title{Format Balance Table}
\usage{
.format_balance_table(
  matching_results,
  round_digits,
  table_style,
  variable_labels
)
}
\description{
Format Balance Table
}
\keyword{internal}
