% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_reporting.R
\name{.format_descriptive_table}
\alias{.format_descriptive_table}
\title{Format Descriptive Statistics Table}
\usage{
.format_descriptive_table(
  descriptive_stats,
  round_digits,
  table_style,
  variable_labels
)
}
\description{
Format Descriptive Statistics Table
}
\keyword{internal}
