% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_reporting.R
\name{.format_robustness_table}
\alias{.format_robustness_table}
\title{Format Robustness Table}
\usage{
.format_robustness_table(
  robustness_checks,
  significance_levels,
  round_digits,
  table_style
)
}
\description{
Format Robustness Table
}
\keyword{internal}
