% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_rdd.R
\name{.perform_density_test}
\alias{.perform_density_test}
\title{Perform Density Test}
\usage{
.perform_density_test(data, running_variable, cutoff, method, bin_size)
}
\description{
Perform Density Test
}
\keyword{internal}
