% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_rdd.R
\name{.prepare_density_plot_data}
\alias{.prepare_density_plot_data}
\title{Prepare Density Plot Data}
\usage{
.prepare_density_plot_data(data, running_variable, cutoff, bandwidth)
}
\description{
Prepare Density Plot Data
}
\keyword{internal}
