% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_rdd.R
\name{.select_rdd_bandwidth}
\alias{.select_rdd_bandwidth}
\title{Select RDD Bandwidth}
\usage{
.select_rdd_bandwidth(
  dt,
  outcome,
  running_variable,
  method,
  kernel,
  polynomial_order
)
}
\description{
Select RDD Bandwidth
}
\keyword{internal}
