% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggraph_transitions.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing Operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Left-hand side value}

\item{y}{Right-hand side default value}
}
\value{
x if not NULL, otherwise y
}
\description{
Internal utility operator similar to %||% in other packages.
}
\keyword{internal}
