% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecshift_class.R
\name{is_vecshift_result}
\alias{is_vecshift_result}
\title{Test whether an object is a vecshift_result}
\usage{
is_vecshift_result(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
A logical scalar: `TRUE` if `x` inherits from `vecshift_result`,
  `FALSE` otherwise.
}
\description{
Test whether an object is a vecshift_result
}
\examples{
dt <- data.table::data.table(
  id = 1L, cf = "A",
  inizio = as.Date("2023-01-01"),
  fine = as.Date("2023-12-31"),
  prior = 1
)
res <- vecshift(dt)
is_vecshift_result(res)
}
