% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_matching.R
\name{print.balance_assessment}
\alias{print.balance_assessment}
\title{Print method for balance assessment}
\usage{
\method{print}{balance_assessment}(x, ...)
}
\arguments{
\item{x}{A balance assessment object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for balance assessment
}
