% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_estimation.R
\name{print.did_results}
\alias{print.did_results}
\title{Print method for DiD results}
\usage{
\method{print}{did_results}(x, ...)
}
\arguments{
\item{x}{A did_results object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for DiD results
}
