% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecshift_modular.R
\name{print.employment_validation}
\alias{print.employment_validation}
\title{Print Employment Data Validation Results}
\usage{
\method{print}{employment_validation}(x, ...)
}
\arguments{
\item{x}{An employment_validation object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Employment Data Validation Results
}
