% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_matching.R
\name{print.matching_diagnostics}
\alias{print.matching_diagnostics}
\title{Print method for data quality diagnostics}
\usage{
\method{print}{matching_diagnostics}(x, ...)
}
\arguments{
\item{x}{A data quality diagnostics object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for data quality diagnostics
}
