% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecshift_class.R
\name{print.summary.vecshift_result}
\alias{print.summary.vecshift_result}
\title{Print method for summary.vecshift_result}
\usage{
\method{print}{summary.vecshift_result}(x, ...)
}
\arguments{
\item{x}{A `summary.vecshift_result` object.}

\item{...}{Unused.}
}
\value{
`x`, invisibly.
}
\description{
Print method for summary.vecshift_result
}
