% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_evaluation.R
\name{print.treatment_event_assessment}
\alias{print.treatment_event_assessment}
\title{Print method for treatment event assessment}
\usage{
\method{print}{treatment_event_assessment}(x, ...)
}
\arguments{
\item{x}{A treatment_event_assessment object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for treatment event assessment
}
