% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecshift_integrated.R
\name{print.vecshift_integrated_result}
\alias{print.vecshift_integrated_result}
\title{Print Vecshift Integrated Results}
\usage{
\method{print}{vecshift_integrated_result}(x, ...)
}
\arguments{
\item{x}{A vecshift_integrated_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Vecshift Integrated Results
}
