% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecshift_class.R
\name{print.vecshift_result}
\alias{print.vecshift_result}
\title{Print method for vecshift_result}
\usage{
\method{print}{vecshift_result}(x, ...)
}
\arguments{
\item{x}{A `vecshift_result` object.}

\item{...}{Further arguments passed to the underlying data.table print
method.}
}
\value{
`x`, invisibly.
}
\description{
Print method for vecshift_result
}
