% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_input.R
\name{print.vecshift_validation}
\alias{print.vecshift_validation}
\title{Print method for vecshift_validation}
\usage{
\method{print}{vecshift_validation}(x, ...)
}
\arguments{
\item{x}{A `vecshift_validation` object.}

\item{...}{Unused.}
}
\value{
`x`, invisibly.
}
\description{
Print method for vecshift_validation
}
