% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_estimation.R
\name{synthetic_control_method}
\alias{synthetic_control_method}
\title{Synthetic Control Method Wrapper}
\usage{
synthetic_control_method(
  data,
  outcome_var,
  unit_var,
  time_var,
  treatment_var,
  treatment_time,
  covariates = NULL,
  method = "augsynth",
  inference = TRUE,
  n_lags = 0
)
}
\arguments{
\item{data}{A data.table containing panel data}

\item{outcome_var}{Character. Outcome variable name}

\item{unit_var}{Character. Unit identifier variable}

\item{time_var}{Character. Time variable}

\item{treatment_var}{Character. Treatment indicator}

\item{treatment_time}{Integer. Time period when treatment begins}

\item{covariates}{Character vector. Time-varying covariates. Default: NULL}

\item{method}{Character. Augsynth method: "augsynth", "synth", "ridge". Default: "augsynth"}

\item{inference}{Logical. Perform inference using conformal method? Default: TRUE}

\item{n_lags}{Integer. Number of outcome lags to include. Default: 0}
}
\value{
A list containing:
  \item{synth_estimates}{Synthetic control estimates}
  \item{unit_weights}{Synthetic control unit weights}
  \item{balance_table}{Pre-treatment balance}
  \item{inference_results}{Inference results if requested}
  \item{plot_data}{Data for synthetic control plots}
}
\description{
Wrapper function for synthetic control estimation using the augsynth package
for cases where traditional matching is not feasible.
}
\examples{
\dontrun{
synth_results <- synthetic_control_method(
  data = panel_data,
  outcome_var = "employment_rate",
  unit_var = "region_id",
  time_var = "year",
  treatment_time = 2015,
  method = "augsynth"
)
}

}
